complete pin authentication process
