--- app/routes/authenticate.js
@@ -1,5 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Route.extend({
-
-});
