+++ app/templates/application.hbs
@@ -1,3 +1,3 @@
-﻿<div>{{#link-to 'index'}}Home{{/link-to}} | {{#link-to 'offers'}}All Offers{{/link-to}}</div>
+﻿<div>{{#link-to 'index'}}Home{{/link-to}} | {{#link-to 'offers'}}All Offers{{/link-to}} | {{#link-to 'register'}}Register{{/link-to}}</div>
 
 {{outlet}}
