+++ app/templates/authenticate.hbs
@@ -1,0 +1,13 @@
+
+<h3> Input Code </h3>
+
+<p>Input 6-digit SMS code we just sent you:</p>
+
+{{input type="text" value=pin name="pin" id="pin"}}
+<button {{action "authenticateUser"}}>OK</button>
+<div class="auth_error">Sorry! Please enter the correct pin.</div>
+
+<p>Haven't received SMS code?<br/>
+Please wait 5 minutes...</p>
+
+<button {{action ""}}>Resend</button>
